cycleMatch: fix 'c'/'p' confusion on video's first frame
It was impossible to set the first frame to 'c', and instead
it was possible to set it to 'p', which doesn't work.

This function doesn't seem to be actually used at all,
but as long as it does exist, it may as well be correct.

Regression in commit e7e797f835456c303bb49c14a096eab1c9863282.